change: sequential take closure to judge the res
