+++ src/sequential/variants/random.rs
@@ -8,6 +8,8 @@ use crate::{Node, NodeGen, NodeResult};
 use crate::nullable_access::NullableWorldAccess;
 use crate::sequential::{ScoredSequence, NodeScorer, };
 
+use super::last_result;
+
 
 /// Weighted random sampling.
 /// Probability of being picked next is proportional to the score.
@@ -53,7 +55,7 @@ impl RandomOrderedSequentialAnd
             node_scorers,
             move |nodes| pick_random_sorted(nodes, (&mut rng.lock().unwrap()).deref_mut()),
             |res| res==NodeResult::Success,
-            NodeResult::Success,
+            |_| NodeResult::Success,
         )})
     }
 }
@@ -78,7 +80,7 @@ impl RandomOrderedSequentialOr
             node_scorers,
             move |nodes| pick_random_sorted(nodes, (&mut rng.lock().unwrap()).deref_mut()),
             |res| res==NodeResult::Failure,
-            NodeResult::Failure,
+            last_result,
         )})
     }
 }
@@ -102,7 +104,7 @@ impl RandomOrderedForcedSequence
             node_scorers,
             move |nodes| pick_random_sorted(nodes, (&mut rng.lock().unwrap()).deref_mut()),
             |_| true,
-            NodeResult::Success,
+            last_result,
         )})
     }
 }
@@ -125,7 +127,7 @@ impl RandomForcedSelector {
             node_scorers,
             move |nodes| pick_random_one(nodes, (&mut rng.lock().unwrap()).deref_mut()),
             |_| false,
-            NodeResult::Failure,  // Be used only when the nodes is empty.
+            |_| NodeResult::Failure,  // Be used only when the nodes is empty.
         )})
     }
 }
@@ -135,7 +137,7 @@ mod tests {
     use crate::*;
     use crate::task::*;
     use crate::tester_util::{TesterPlugin, TesterTask, TestLog, TestLogEntry};
-    use crate::sequential::{NodeScorerImpl, ConstantScorer};
+    use crate::sequential::{NodeScorerImpl, variants::ConstantScorer};
     use super::*;
 
     use rand::SeedableRng;
